Move headers to top-level include directory and update build configuration
