+++ include/Accounts/SetUserFromDB.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../pch.h"
+#include "pch.h"
 #include "Accounts/User.h"
 #include "Accounts/AccountManager.h"
 #include "System/Database/MongoDBDataManager.h"
