+++ include/Accounts/UserBusiness/BusinessRepository.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../../pch.h"
+#include "pch.h"
 #include "Accounts/User.h"
 #include "Accounts/AccountManager.h"
 class BusinessRepository {
