+++ include/Accounts/UserBusiness/SetBusinessFromDB.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../pch.h"
+#include "pch.h"
 #include "Accounts/UserBusiness/BusinessRepository.h"
 
 class SetBusiness {
