+++ include/Stock/ClientStockItem.h
