+++ include/Stock/StockDetails.h
