+++ include/Stock/StockItem.h
