+++ include/Stock/StockManager.h
